[fix] add confirm dialog delete property (#1458)
ESP:

![image](https://github.com/onparallel/parallel/assets/11854260/9ed271a5-86e9-4f3b-86e3-40d35e1b7a42)

![image](https://github.com/onparallel/parallel/assets/11854260/37881500-31cc-4e70-b4cc-a6c2e3cd9ce6)

ENG:

![image](https://github.com/onparallel/parallel/assets/11854260/c7706586-b202-4ce3-9023-d39bb49d33fe)

![image](https://github.com/onparallel/parallel/assets/11854260/050a5501-2826-40e4-8132-6558af9d2893)

+++ server/src/graphql/profile/mutations.ts
@@ -356,7 +356,7 @@ export const deleteProfileTypeField = mutationField("deleteProfileTypeField", {
       })
     ),
   },
-  resolve: async (_, { profileTypeId, profileTypeFieldIds }, ctx) => {
+  resolve: async (_, { profileTypeId, profileTypeFieldIds, force }, ctx) => {
     const profileType = (await ctx.profiles.loadProfileType(profileTypeId))!;
     if (
       profileTypeFieldIds.some((id) =>
@@ -368,9 +368,19 @@ export const deleteProfileTypeField = mutationField("deleteProfileTypeField", {
         "FIELD_USED_IN_PATTERN"
       );
     }
+    const profileCount = await ctx.profiles.countProfilesWithValuesOrFilesByProfileTypeFieldId(
+      profileTypeFieldIds
+    );
 
-    // TODO: Check if field has value or files attached to it
-    // if does, return error FIELD_HAS_VALUE_OR_FILES
+    if (Number(profileCount) > 0 && !force) {
+      throw new ApolloError(
+        "At least one of the provided profile type field ids has value or files.",
+        "FIELD_HAS_VALUE_OR_FILES",
+        {
+          profileCount,
+        }
+      );
+    }
 
     await ctx.profiles.deleteProfileTypeFields(
       profileTypeId,
